assignment3/requirements.txt
google-api-python-client==2.103.0
google-cloud-storage==2.11.0
flask==2.3.*
functions-framework==3.*
google-cloud-logging==3.8.0
google-cloud-pubsub==2.18.4